requirements.txt
streamlit
google-genai
pillow
pypdf
pdf2image
gTTS
fpdf2